Fix "find expensive includes" algorithm and add initial tests
The original algorithm didn't work and this was discovered in the newly
added `find_expensive_includes.test.cpp`.

The new algorithm does 2 depth-first searches, one from the source file
and one from the file who is no longer included.  We find all files that
were found in the second search but not the first and sum their file
sizes.

This may invalidate the `reachability_graph` component, but this will be
left for the moment.

+++ src/build_graph.cpp
@@ -40,7 +40,7 @@ class IncludeScanner : public clang::PPCallbacks {
           std::filesystem::path(file->getName().str()).lexically_relative("./");
       return m_lookup
           .emplace(file->getUID(),
-                   add_vertex({p.string(), file->getSize()}, m_graph))
+                   add_vertex({p.string(), static_cast<std::size_t>(file->getSize())}, m_graph))
           .first->second;
     }
   }
